refactor and remove heavy-weight test
