+++ tests/sparseml/export/transformers/test_transformers.py
@@ -28,7 +28,6 @@ from sparsezoo import Model
     "stub, task",
     [
         ("zoo:obert-medium-squad_wikipedia_bookcorpus-pruned95_quantized", "qa"),
-        # ("zoo:roberta-large-squad_v2_wikipedia_bookcorpus-base", "qa"),
     ],
 )
 class TestEndToEndExport:
@@ -37,7 +36,6 @@ class TestEndToEndExport:
         model_path = tmp_path / "model"
         target_path = tmp_path / "target"
         self.model = Model(stub, model_path)
-        self.is_model_quantized = stub.endswith("quantized")
         source_path = self.model.training.path
 
         yield source_path, target_path, task
@@ -112,11 +110,6 @@ class TestEndToEndExport:
         )
 
     def test_export_validate_correctness(self, caplog, setup):
-        if self.is_model_quantized:
-            pytest.skip(
-                "Skipping since quantized models may not pass this test"
-                "due to differences in rounding between quant ops in PyTorch and ONNX"
-            )
         source_path, target_path, task = setup
 
         num_samples = 3
